HEEDLS-541 move new classes into new file
